Finalizacion

+++ Models/Student.cs
@@ -9,6 +9,6 @@ namespace AP1.Models
         public int FirstPartial { get; set; }
         public int SecondPartial { get; set; }
         public int Systematic { get; set; }
-        public int FinalNote { get; set; }
+        public double FinalNote { get; set; }
     }
 }
